Fix label handling for perturbed graphs
This is quite a big change.  The labels used to be handled using "usual
arithmetic" or, as was pointed out, "some rounding between friends."

Now we slap the value of the edge directly onto the label's memory using
`memcpy`.  Much cleaner.

Added the tooling to do that in 'cpd/cast.h'.

Maybe we will need to make sure we are not writing beyond the memory
allocated to the label at some point.