Change build process
Now split into two: fast/dev, and build in their respective directory.
This is achieved by having a *base Makefile* which handles creating
variables (sources, compile options) and creating a Makefile in the
target directory.

Then compilation is as before. Note that this can only be invoked from
`src/` as the paths are relative.